Do not cache S3 connections
Caching S3 connections using the S3 endpoint as the key to a map of cached S3
connections is incorrect, because, when other aspects of an S3 connection like
its secret, region, etc., changes, a new S3 connection is not formed.  Hence,
do not cache S3 connections (a premature optimization that is not proven as
required).

Signed-off-by: Veera Deenadhayalan <vdeenadh@redhat.com>