Bug 2059303: Fix static strings in config map append
The bundle overlay for odr contained strings that were
already odr specific, which hence were not replaced by
the Makefile when building the bundle.

This commit rectifies this by changing the defaults to
ones that the Makefile would replace, hence fixing the
incorrect clusterServiceVersionName in the config map
that was bundled.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>