tests: use the FakeObjectStore in the unit tests
Earlier, we were using the same object as the real object store in the
tests. Using the FakeObjectStore would allow us to simulate the success
and failure cases with the object store.

The FakeObjectStore was already implemented and used by some of the
drpolicy tests; this commit makes the vrg tests to also use it.

Co-Authored-by: Raghavendra M <raghavendra@redhat.com>
Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>
(cherry picked from commit 223a21261d5e0bf8f8133df1687d09ef91455633)