Bug 2085357: Change DRCluster resource to a cluster scoped resource
A DRCluster represents a 1:1 relationship to a managed
cluster, and hence for reasons such as fencing, we need
to maintain a single instance of the same across the hub
control plane as well.

Moving this to a cluster scoped resource as a result.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 9fb190156253663f1b0ed0bde9fa12f0ec483f05)