update drpc status when vrg status condition changes
If nothing has changed in the DRPC, then the status is not updated. The DRPC status contains
the VRG status conditions but it was't checking whether the VRG status condition has changed.
In this PR, and if the VRG status condition has changed, then we update the DPRC status to
reflect the change.

(cherry picked from commit 27e206efea44b628e0e274bfbc6c4d8798cff646)