Set VR autoresync based on VRG spec
- Set to true only when moving to secondary and
due to failover

In all other cases of moving to secondary it is a
relocate, and hence a known primary is being demoted
and does not require a auto resync.

If being set as the primary, then there is no resync
operation to control anyway.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 2a1dfad78513225726806cc5be625c6f9e20694b)