Update reconciler to requeue for specific cases
Current VRG requeue was generic, deciding on both
desired status of VR and on other failures.

This change separates the 2, and only requeues for
cases where it is required to reconcile as there would
be no external event to trigger a reconcile. The other
case where VR has not reached desired state is bubbled
up the stack for decisions, but not a reconcile requeue
case, as that would occur when VR updates itself.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 003c7165d97dedd4fe8acf0aca5b0d7e7552566b)