Update downstream main to contain DS artifacts
Each branch of downstream main causes additional
changes to be submitted to the release branch.

This commit removes the need to do so, and instead
adds the required changes to downstream main branch.

When rebasing downstrem main to upstream main, it may
result in conflicts that need to be resolves during the
rebase. It is expected that other than sanuty.yml no other
file/content should at present be in conflict.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>