s3utils: add deadline to s3 operations
When a s3 store is coming up or going down, it might not respond to a
request. This change in code adds a deadline to the operations to
prevent the main thread from being locked on such calls.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>
(cherry picked from commit 8f4ffe185a20d0321329bf0928552db3f53c4c99)