MetroDR: clear PV ClaimRef on secondary PVC deletion and put back the ClaimRef on restore
In MetroDR we don't change the reclaim policy for the PV back to  when we
failover/relocate an app to another cluster. In other words, if we failover from
C2 to C1, then on C2, we will not change the reclaim policy back to .
Therefore, the PV will not be deleted but its Status.Phase will move to Released.

In that case, if we failover or relocate back to that cluster, then it is guaranteed
that a new PV will be created instead of reusing the Released one.

Kubernetes will only bind a PVC to a PV if the PV is in the Available status phase.

The fix in this PR we are removing the ClaimRef from the PV when the PVC is deleted.
On Restore, if we find that the PV exists, then we put back the original ClaimRef from
the s3 store.

(cherry picked from commit c6758ae09b714eb1d63857e644f365f9bd94c692)