Address missing VR for a PVC during VRG deletion
There could be cases where a PVC finalizer is present
but, during VRG deletion, post deletion of the VR the update
to remove the finalizer on the PVC failed.

This would on the next reconcile recreate the VR, which stalls
deletion processing as the VR cannot add finalizers to the PVC
and hence cause the deletion to loop forever and fail.

This commit adds checks to ensure VR is missing and if so
complete the deletion process for the PVC, and if not perform
VR deletion orchestration as usual.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 01d4e722c466d6a6949f5dd021a20200c648f6d9)