Reword lastSyncTime to lastGroupSyncTime
In DRPC and VRG, lastSyncTime is for a group of PVCs
and not for a single PVC. As this leads to some confusion
it is reworded as such.

Per PVC still would report lastSyncTime, as is the case
with Volsync and VolumeReplication resources.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 242daf2917e769e2fa98e15c0e3e8e7374586909)