Ensure placement ref in DRPC is in the same namespace
PlacementRef will be used to update the PlacementRule as per
the reference provided. If it is in a different namespace
than the DRPC, it would potentially be a vulnerability.

This commit checks if the namespace is the same, if not
errors out withough reconciling the DRPC resource or updating
the placement rule.

Fixes: #535

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 0da26e637fbf899d3c152fb717410d4c520bba81)