Update makefile to disable KubeObjectProtection
bundle builds will now disable kubeObjectProtection in
the ramen config file for both the hub and the dr-cluster.

Auto configuration of the dr-cluster would hence adopt the
hub configuration which is disabled.

A cleaner default would be true, with the build team disabling
the same in their builds, but that can be done when the
feature nears maturity.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>