Replace PVC in use check with VA and pods using PVC check
Current check to ensure PVC is not in use, specifically when
VRG is moved to Secondary for relocation, relied on the PVC
in use finalizer. This finalizer is removed earlier than the
backing PV being unmounted/unmapped from a node. As a result
in relocation cases this can cause data corruption as cached
data maybe written post demoting the volume to Secondary.

Instead of this the current code reuses the volsync method to,
- Ensure there are no pods referencing the PVC
- Ensure there are no pending VolumeAttachments for the PV

The above is checked post PVC deletion, which results in no
further possibility for the PVC to be actively used before
demotion. Which hence avoids the potential corruption.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 3fdc1be04344992abca567d8c8fa25d8d943a699)