Rebuild DRPC state when DRPC is recovered from backup
Two of the issues that we've seen with the hub recovery tests are:
1. VRG MW didn't get backed up
2. Velero not restoring the resources with their status.

For the first issue, the fix is easy; just add the backup label to the VRG MW.
For the second one, the PlacementRule depends on the DRPC to update its status.
That didn't work when the action has already been changed to 'Failover' or 'Relocate'
prior or at the time of the resource backup.

This PR fixes the first issue by adding a backup label to the VRG MW. And it fixes
the second issue by querying the Managed Clusters for VRGs and rebuilds the DRPC
state with the query result.

If the DRPC state can't be rebuilt, for example, let's look at the following scenario:
1. The user failed over from C1 to C2
2. Hub backup is initiated
3. The user decides to relocate back to C1 from C2
4. Before another hub backup is taken, the hub crashes and the user needs to recover
it somewhere else.

In that case, DRPC detects that the last action was a 'Failover' (from the drpc.spec.action),
but it finds the primary VRG in the non failover cluster, which is contradicting. In that case,
the user has to intervene and corrects the drpc.spec.action manually buy setting it to 'Relocate'.

Signed-off-by: Benamar Mekhissi <bmekhiss@redhat.com>
(cherry picked from commit ea569466d359489fea5c77953e09286de50d8220)