Use CopyMethod Direct when setting up VolSync ReplicationDestination
In this PR, added the option for VolSync to sync directly to the application PVC.
This will resolve the issue where CephFS takes a long time to create an RWX PVC
from a snapshot. Moreover, it resolves the issue where in some cases (PVC data size and latency)
the first sync after failover/relocation takes a long time to fully copy the PVC to the
destination side. Instead, with this change, only the delta is transferred.

By default, the CopyMethod from source to destination will remain snapshot-based.
If users want to use this new feature, they have to enable it through the Ramen configmap.
```
volSync:
  destinationCopyMethod: "Direct"
```
The caveat with this solution is that (affects only the failover), there is a probability of
about 5% where the delta sync is in progress when we lose connectivity to the secondary cluster.
If the user decides to failover at that point, then the data written to the destination PVC
might be inconsistent as opposed to crash-consistent.

Signed-off-by: Benamar Mekhissi <bmekhiss@redhat.com>
(cherry picked from commit 1a6cd24527d67ef1adf7f019e1a04c4d473d87a2)