Cleanup PV claim ref. during restore and not delete
Current code cleans up the PV claim ref. if metro DR is
in use, at the time of deleting VRG on a cluster.

If the PVC is not deleted as part of the workload cleanup
then due to the PV claim ref. being cleaned, the PVC reports
its state as "Lost".

This is observed due to changes made as part of commit
cf63370e3fd8025f1a55d64cc6880a3941b958fc which started
supporting failover in cases where PVCs are not deleted.

The fix is to not perform the PV claim ref. cleanup on delete
and instead continue to cleanup the PV claim ref. when the PV
is restored to the cluster as part of validatePVExistence.

In validatePVExistence if the pvExists and matches the to be
restored PV, and is Bound, it is ignored. If not, it is
cleaned up subsequently by updateExistingPVForSync for metro
DR cases. updateExistingPVForSync will ensure a recreated PVC
will bind back to the PV, and pvExists would ensure that an
already Bound PV is left as is.

Fixes: #652
Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit e73ff3633fec6510a94c806d774e71ff72bbfc1a)