Use peer cluster as the current homeCluster when it cannot be discovered
During a failover, Ramen conducts its own validation and attempts to locate the
cluster where the application was running. However, when restoring DRPC from backup,
the status will not be restored with it. Ramen must discover where the app was
running in order to rebuild it state. In the case of a double failure, where both
the hub and the managed cluster are down, Ramen cannot determine where the
application was running and fails to proceed. This commit defaults to the peer
cluster as the current homeCluster before failover.

Signed-off-by: Benamar Mekhissi <bmekhiss@redhat.com>
(cherry picked from commit 00c4e05ad33b95444ed5a54aa0fc4f421244118f)