Update VRG to v1alpha2 version
Create a space for v1alpha1 to alpha2 upgrade,
as IBM fusion 2.4 has released an alpha1 CRD that
is slightly different than current alpha1.

This commit hence unconditionally moves VRG resource
to alpha2, thereby creating space for a webhook based
version conversion oppertunity for the Fusion product
stream in 2.5, and in ramen upstream.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>