Ordered initialization for the DRPC condition array
Introduce ordered initialization for the condition array in the DRPC module.
This change ensures that the conditions are consistently located in the array,
allowing for easier management and maintenance. Additionally, the updates made
to the code now properly preserve the order of the conditions, providing more
reliable and predictable behavior.

Signed-off-by: Benamar Mekhissi <bmekhiss@ibm.com>
(cherry picked from commit e9a78fc6001cca3a77e334e39657475868a0f483)