Disable vrgNamespace caching to prevent upgrade issues
Previously, we implemented a caching mechanism for vrgNamespace to optimize
the process of fetching the namespace of the application when it is part of
the ApplicationSet. However, this approach caused an upgrade problem when
moving from version 4.12 to 4.13. This commit will disable the vrgNamespace
caching until all customers have upgraded to version 4.13 or later.

Signed-off-by: Benamar Mekhissi <bmekhiss@ibm.com>
(cherry picked from commit 572c9dcbfb9cefe39adffd26e369baf22e335a1b)