handle error when AppSets CRDs are missing
We see an error when trying to list AppSets when
AppSet CRDs are not installed hence ramen was in
an error state and unable to proceed as the namespace
was not returned. This is fixed by returning the default
namespace when AppSets CRDs are not found.

Bug: https://bugzilla.redhat.com/show_bug.cgi?id=2218181

Signed-off-by: rakeshgm <rakeshgm@redhat.com>
(cherry picked from commit 098592209cecc54d62cc06499c507da048c0488d)