Update hub RBAC to include Placement finalizer "update"
As per [1] when setting owner references to a resource and
specifically while setting blockOwnerDeletion to true, an
RBAC to update the finalizer of the resource set as the owner
is required.

This commit adds the required RBAC for the same.

[1] https://kubernetes.io/docs/reference/access-authn-authz/admission-controllers/#ownerreferencespermissionenforcement

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 11b77299c2caab78cb37c904218015de68906de2)