Propagate OCM annotations from primary PVC
To ensure that OCM owns the volsync protected PVC when the application
is deleted or DR is disabled, copy the OCM annotations from the PVC to
the protectedPVC. The annotations are propagated to the target clusters
and used to create the PVC during failover and relocate.

Tested by adding a new section in existing test. The test ensures that
we copy only the OCM annotations and drop other annotations.

Fixes: #1033

Signed-off-by: Nir Soffer <nsoffer@redhat.com>
(cherry picked from commit f19c8ed437f8f66f39d9c6611145ffd650c187fc)