Add retry attempts in case of PlacementDecision name conflicts
As we create a PlacementDecision if we do not find one, it could
lead to conflicts on the name. This commit adds a retry with different
index values for the name, with a cap count of 5, to ensure some
defence against conflicts.

An ideal case would be to just use a uuid as the index, and avoid
conflicts across, but that needs some confirmation with OCM on the
PlacementDecision naming scheme and hence is not done at present.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit f9514cc05b562bf8038b2bb672306c9a54c9b98b)