Filter out the do-no-delete annotation
This annotation is added to the application PVC on the source relocate.
At the time the annotation is added the protectedPVCs were already
propagated to the target clusters, so it is never propagated to the
target clusters. However if the annotation will be propagated to the
target clusters it can break things, so now we filter it out when
copying the PVC annotation to the protected PVC.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>
(cherry picked from commit 825a0de23b8f149a444aebfc2b5ed74454fc8e5b)