Automatically generate secret for the test VM
We can generate the secret during kustomization, but it is not possible
to refer to `~/.ssh/id_rsa.pub`. We cannot use a static symlink, and if
we create a symlink we need to use

    kustomize build --load-restrictor=LoadRestrictionsNone

which makes it harder to apply the kustomization manually for testing.

To keep things simple, we copy we copy the user default public key to
the vm kustomization directory during the test. From this point the
kustomization can be applied manually.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>