Pass in the current primary to createVolSyncDestManifestWork during cleanup
On hub recovery cases when the ManifestWork does not exist on the hub
cluster, it is recreated using createVolSyncDestManifestWork.

The function creates VRG for volsync destination on clusters that need
to be Secondary, and has to be passed the primary cluster.

This fix addresses the above during cleanup where the cluster passed was
not the primary.

This was introduced as part of commit:
ea0c1e41e2a20917d9a84a0d12d73e6f27fc069e

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit d2534525c58047a9b482fc1b585ab5474c80dba0)