Add drClusterIsDeleted() helper
In cluster replacement flow, we may need to disable DR when a drcluster
is not available, which breaks disable DR flow in many places.

To allow cleanup to proceed on the hub and on the remaining managed
cluster, the user will have to delete the drcluster for the failed
cluster.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>