Selectively define ApplicationSet spec to drop argocd mod deps
*** THIS IS A HACK!!!

To remove the k8s and subsequent go mod dependency problems,
selectively define a local ArgoCD ApplicationSet type within
the code, and use the same for listing and looking at required
fields in an ApplicationSet.

If ArgoCD API versions change, we may have to relook at the
fields and which are non-optional, to redo parts of this for
the next API version.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit d0abb28ce29549c896b9997850bc91439e7b8b93)