Disown volsync managed pvcs upon vrg deletion
This commit introduces changes to disown PVC resources when a VRG is deleted,
particularly in scenarios where the DR is disabled. The motivation behind this
adjustment is to give back the management of the PVC lifecycle to OCM.

Upon disabling DR, the VRG now iterates through all volsync-managed PVCs.
It removes VRG ownership from these PVCs and reinstates OCM annotations.
This ensures that PVCs are not prematurely garbage collected upon VRG deletion.

In order for that to work, a new annotation needs to be added to the VRG to
indicate that the PVCs should be disowned upon DR disabling
drplacementcontrol.ramendr.openshift.io/do-not-delete-pvc: "true"

Signed-off-by: Benamar Mekhissi <bmekhiss@ibm.com>
(cherry picked from commit d82d15511cca9adf2a2ae653cb6fdc00aaefd26a)