Check for metro policy using policy clusters
Current check for metro actions that use isMetroAction passed
in the to and from clusters to check equivalence of region
values. This would succeed when to and from clusters are the
same, which can happen when the preferredCluster and the
failoverCluster are the same (Deploy->Failover->Failover).

As a result in a regional DR policy, such a state would enter
processing or waiting for fencing to take effect before
proceeding with the Failover.

This commit switches to the policy based metro determination
instead of the to and from clusters as before to resolve the
issue.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit b76fe12817e5d4d64ab2c41555b641be139ec65a)