go.mod: create a new module for the api package
This commit adds the go.mod and go.sum files for the api package to make
it an independent module. The go.mod file at the top dir now references
this new module with a replace directive.

In the next PR, we will remove the replace line and add the dependency
using the latest commit.

Essentially a backport of e11c56eac24d2992938e0120663bf7755689155c

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>