Add DRPC UID as an annotation to VRG
This helps determine using an MCV later if the VRG on the managed
cluster was created by a manifest work from the current DRPC
resource instance.

In cases of hub recovery this will mismatch as the DRPC is
recreated on the hub and hence would have a different UID and
helps to not orphan VRGs on the managed clusters due to missing
hub manifest work for the same.

The change to add the UID in the initial manifest work creation
ensures other parts of the code that fetch the manifest and update
it are not impacted. IOW, if a manifest work exists for a VRG it
is created by the current instance of the DRPC as manifest work
resources are not recovered on a hub recovery or in a declarative
creation of the DRPC.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 70fcd9a1b1e35c724e19dc2abdbe652d311f4df1)