Adopt any orphan VRGs found via views before starting reconciliation
Adoption ensures that there is a ManifestWork that manages the VRG
resource on the ManagedCluster.

In the case of hub recovery, we would find VRGs on the managed cluster
without any ManifestWork for the same on the hub. This change ensures
that such missing ManifestWork is created on the hub by the current
DRPC resource.

It does this by storing the DRPC resource UID as an annotation on the
created VRG, and on any discrepency ensures that the current DRPC creates
a manifest work for the VRG on the new hub.

The case for upgrade is also handled, where a manifest work may be present
but without the DRPC UID annotation on it.

Merge conflicts:
- ResourceIsDeleted function is not present in current branch, hence
code using this is updated to check deletion of resource as before

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 15c8418872b24910021eeafca4fe46629c63eca4)