Ensure cleanup of policy and secrets is complete...
When a corresponding PlacementRule for the Policy is not
found during removal of a secret.

It can so happen that the PlacementRule is deleted but the
finalizer on the secret is not removed, at which point if we
do not ensure cleanup the secret may retain a stale finalizer.

While it is possible to only selectively remove the finalizer,
ensuring all other resources are cleaned up as well helps, hence
the delete routine is invoked, than just checking and removing
the finalizer on the secret.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 32ac9065b850e6296e362e8b0183aa271d5b6c84)