Using S3 Storage to Recover lastGroupSyncTime After Hub Recovery
After a Hub recovery, the DRPC status is not restored, resulting in the loss of
the lastGroupSyncTime information, which cannot be reconstructed if the primary
cluster is also down. To address this issue, use an S3 store to recover the lost
information.

Bug: https://bugzilla.redhat.com/show_bug.cgi?id=2275320

Signed-off-by: Benamar Mekhissi <bmekhiss@ibm.com>
(cherry picked from commit 6379427e4c047b63befe88b282df0c9ca7cb8017)