Default to ramen created velero secret, if ref is missing in config
If veleroNamespaceSecretKeyRef is empty/nil in the ramen config
default to using the ramen orchestrated velero secret ref.

This ensures that the S3 secret that is exchanged with the managed
cluster in the velero namespace is used for required velero
operations.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit b6a072bd03af6b59980d4905b4bf8484df8b85a7)