vrg: get start time of recover/backup from first velero req
Earlier, we assumed that all steps in a workflow sequence are velero
requests. However, it is not true anymore when hooks are allowed in a workflow.
We get the start time of the job from the first velero request.

Co-Authored-by: Annaraya Narasagond <annarayanarasagond@gmail.com>
Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>
(cherry picked from commit 509b2f0d1aeb8af59f81399c19dfc045ba97284c)