vrg: don't protect kube resources as secondary during relocation
This is not the right place to protect the kube resources as one final
sync when we are relocating. It is too late as the user has already been
informed to cleanup the resources and some of the resources might have
been deleted already.

The right time to do it is before the application resource cleanup but
we cannot do it right now because of the difference in how volrep and
volsync behave.

We will fix it in a future release.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>
(cherry picked from commit 32a5bc3b47278d21e65dd16784a034f30d67913f)