Use reason to conditionally restore cluster data
Currently cluster data, PV/PVC and kube objects, are restored if the
generation of the VRG changes. It is skipped only when the generation
changes and the spec contains one of prepare or run final sync.

In cases where the spec can change due to other reasons, as Primary the
resource restore should be skipped if already performed. One such case
is when the peerClasses in VRG are updated with latest information.

This commit fixes the above more generically, by leveraging the condition
reason to trigger a restore when Primary and otherwise ignoring the restore
and updating the condition to the latest generation. To achieve the above
when processing as Secondary as well the reason is reset, such that when
a VRG is converted from Secondary to Primary the reason mismatches and a
restore is performed as desired.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit bae076b0fb560a6c3d14269d5732f55fe832e952)