Add status fields to DRClusterConfig to list available classes
Currently we create OCM ClusterClaims for each class that we detect on
a cluster. This CRD change is to change this to reporting the same in
the status of the DRClusterConfig resource instead.

This is to avoid any limitations on the number of ClusterClaims that may
be enforced by OCM on a cluster.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit 93a4f1a96cb3ee170738a4978c6a99c2259ae008)