pvc: re-upload the pvc to s3 stores when resized
Whenever there is a PVC resize, the PVC was not uploaded to the s3
stores, as a result of which during failover or relocate, the original
size of the PVC was used. This could lead to data loss for the users.

The approach used is to hash the pvc for spec, metadata and objectmeta
barring all the fields which changes often and add as part of value of
archived annotation.

Signed-off-by: Annaraya Narasagond <annarayanarasagond@gmail.com>
(cherry picked from commit 4ef49a2d1da20c742df1d207aa544e8021cae93f)