Ensure do-not-delete-pvc annotation propagates before allowing VRG deletion
When disabling DR, the UI adds the do-not-delete-pvc annotation and then deletes
the DRPC in a second step. However, by the time Ramen reconciles the DRPC resource,
it may already be in a deleted state. In that case, the annotation wouldn’t propagate
to the VRG, which could result in unintended deletion of application PVCs.

This change updates the controller to verify that the annotation is present on both
the DRPC and the VRG before permitting VRG deletion. This ensures the annotation is
properly propagated and prevents accidental PVC removal.

Signed-off-by: Benamar Mekhissi <bmekhiss@ibm.com>
(cherry picked from commit abd24b14c3373d1aa705bca14deca32c1c0cd773)