drpc: add argument to CreateOrUpdateNamespaceManifest
In the mw_util, CreateOrUpdateNamespaceManifest was not accepting labels
as one of the argument and was always sending empty labels. Parameterizing,
so that any consumers of it will be able to send the labels as well now. The
existing callers are made to send empty labels so that they are not impacted.

Signed-off-by: Annaraya Narasagond <annarayanarasagond@gmail.com>
(cherry picked from commit 9ec3b194da706bef114ea47131d4bb1bd12dcc06)